requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
scikit-learn==1.7.1
pandas==2.3.1
joblib==1.5.1
requests==2.32.4
streamlit==1.47.1
httpx==0.28.1
python-dotenv==1.1.1
